requirements.txt
pandas
matplotlib
python-pptx
fpdf